refactor(config): Rename fast_api module to fastapi across config versions
- Rename `fast_api.py` to `fastapi.py` in v0 and v1 config modules
- Update import statements to reflect the new module name
- Maintain existing FastAPI configuration structure
