+++ cogito/core/config/v0/__init__.py
@@ -1,8 +1,8 @@
 """V0 configuration models"""
 
 from cogito.core.config.v0.base import CogitoConfig
-from cogito.core.config.v0.fast_api import FastAPIConfig
+from cogito.core.config.v0.fastapi import FastAPIConfig
 from cogito.core.config.v0.route import RouteConfig
 from cogito.core.config.v0.server import ServerConfig
 
-__all__ = ["CogitoConfig", "FastAPIConfig", "RouteConfig", "ServerConfig"] 
+__all__ = ["CogitoConfig", "FastAPIConfig", "RouteConfig", "ServerConfig"]
