+++ cogito/core/config/v1/__init__.py
@@ -1,8 +1,8 @@
 """V1 configuration models"""
 
 from cogito.core.config.v1.base import CogitoConfig
-from cogito.core.config.v1.fast_api import FastAPIConfig
+from cogito.core.config.v1.fastapi import FastAPIConfig
 from cogito.core.config.v1.route import RouteConfig
 from cogito.core.config.v1.server import ServerConfig
 
-__all__ = ["CogitoConfig", "FastAPIConfig", "RouteConfig", "ServerConfig"] 
+__all__ = ["CogitoConfig", "FastAPIConfig", "RouteConfig", "ServerConfig"]
