+++ cogito/core/config/v1/fastapi.py
@@ -1,4 +1,4 @@
-from cogito.core.config.v0.fast_api import FastAPIConfig as v0
+from cogito.core.config.v0.fastapi import FastAPIConfig as v0
 
 
 class FastAPIConfig(v0):
